Handle named multiplexers without signals
Our DBCs include messages in which, for some values of the multiplexer
signal, the multiplexer signal is the only signal included in the payload.
In those cases we indicate values that the mux ID may have by including
them in the enumeration.

The DBC entry for such messages might look like this:

```
BO_ 1000 ExampleMessage: 8
 SG_ ExampleMultiplexer M : 0|8@1+ (1,0) [0|0] ""
 SG_ AdditionalSignal m1 : 8:8@1+ (1,0) [0}0] ""

...

BA_ "FieldType" SG_ 1458 ExampleMultiplexer "ExampleMultiplexer";
VAL_ 1000 ExampleMultiplexer 5 "EVENT_TYPE_FIVE" 4 "EVENT_TYPE 4"...;
```

This commit allows us to parse messages in this category without
raising a DecodeError by including the named choice values of multiplexers
in the codec.