Add .PHONY targets to Makefile
Currently the Makefile does not specify a .PHONY target, which causes
make to treat the target as a file. Since the current make targets
{test,test_c,test-sdist,release-to-pypi} do not create these target
files, if a file of the same name exists in the repository the target
will not work correctly.

Adding .PHONY specifies that the target is not a file, such that the
target will run the recipe regardless of whether there is a file of the
same name.