Separate MUX'd signals in monitor output
The way this works is by using each of the Multiplexer Signals to create
a logical "composite key" for the name of any Message that contains
multiplexed data (both Simple and Extended). This composite key is
simply the Message name prepended to all the possible MUX Signals for a
particular message (and their values). This composite key is therefore
unique for all the different possible enumerations of MUX values, which
allows us to display each MUXed Message separately.