ARXML: add support for loading initial signal values
This only supports specifying the numerical value in the I-SIGNAL as
well as a reference to to constant in the I-SIGNAL. There seem to be
quite a few additional methods which remain unsupported. As far as I'm
concerned, this is *way* too complicated in ARXML.

Also note that I do not really have any tools to verify that these
values are specified correctly. I verified that system-4.2.arxml still
validates against the XSD for autosar 4.3, but this seems to only be a
necessary precondition for correctness, not a sufficient one...