ARXML SystemLoader: introduce new ARXML traversal functions
The `follow_arxml_reference()` method follows an arbitrary relative or
absolute ARXML reference to its target node (surprise!), the method
`get_arxml_children()` returns the list of child nodes of a given set
of elements that match a specification. (Note that the syntax of
locations has been slightly extended compared to what was there
before: if a tag name is preceeded by an '&', ARXML references are
allowed, if is prefixed by '*', multiple nodes are possible. These
qualifiers can also be combined.) Finally, there is the
`get_unique_arxml_child()` convenience function which is just like
`get_arxml_children()` but expects to match at most one child node,
i.e., the returned object can be used as before.

Note that the `follow_arxml_reference()` method is basically a
generalized version of the old `SystemLoader.find()` method, whilst
`get_arxml_children()` is an extended version of using ElementTree's
`iterfind()` method on nodes. (It is more general, because it can use
on multiple nodes as its base and it also can resolve ARXML references
automatically.)