sets initial signal value when none supplied during encoding
This is a 2 fold PR because one change is dependent on the other. When encoding a message if a signal was not supplied in the data to be encoded but the signal has an initial value an EncodeError would be raised. This PR changes that behavior to check for an initial value and if one is available it will use that to encode the message with

The second part is to allow encoding a message from the signal directly. The use case for this is if all other signals that are apart of the message have a default value the signal can directly be used to encode a message with the supplied value. It also takes into account if the signal has a multiplexer signal and it if does it sets the value of the multiplexer signal to the name of the signal where encode() was called.

I am sure that there might be a better way to go about handling this and I do not have an in depth knowledge of can databases. In my specific use case I have updated the OBD2.dbc file that is available from
https://www.csselectronics.com/screen/product/obd2-dbc-file/language/en
which in it's downloaded form only decodes messages and does not have the ability to encode messages. I have updated it so it now does have the ability to decode messages and this is done by using initial values and multiplexers. So no actual data needs to be supplied in order to encode an OBD2 message.
if a copy of the updated version is wanted let me know and I will upload it.

use code example
```python
import cantools

db = cantools.db.load_file(r'OBD2.dbc')
message = db.get_message_by_name('OBD2_TX')

signal = message.get_signal_by_name('Odometer')
print(signal.encode(message))

data = {
    'ParameterID_Service01': 'Odometer',
    'Odometer': 0,
    'service': 1,
    'length': 2
}

print(message.encode(data))
```

and the output is exactly the same
```
b'\x02\x01\xa6\x00\x00\x00\x00\x00'
b'\x02\x01\xa6\x00\x00\x00\x00\x00'
```
so in order to transmit the obd2 request the only information that is needed is the name of the message and the name of the signal. Where as the current code makes it so that quite a bit of information needs to be known in order to encode the exact same message.

This is more of a proof of concept then anything else and I am sure that it can be improved upon. my database alterations are within the specification and I am using Vector CANdb++ to make the changes so it is within the ability of the current specification to function in this manner.