all tests pass
I had to update the tests so the file paths would be correct when the tests were run from the inside of an IDE like PyCharm. The way the tests were written originally you MUST run the tests from the parent directory of the tests directory, this is no longer an issue and the tests can now be run without the need for the parent being the cwd.

All of the AssertEquals calls were incorrect the expected value should have been the first parameter and the actual value the second. I have correct the majority of them but I am sure that I probably missed some.

Fixed command line dump of a J1939 database so it properly dumps the arbitration ID. I still need to update this dump to handle GM ids.

When dumping a database to a string using the `as_dbc_string()` method the output is now more organized specifically the attributes. Because of this it required me to update the tests to work with the new output.

There is no proper handling of default values and those values being populated into each objects set of attributes. If the attribute for an object differs from the default it will output the attribute line needed for that object when using `as_dbc_string`.

the attributes are correctly handled for all object types (database, message, signal, node, envvar)

I know that this is going to need some more or it can be used as a proof of concept whichever you prefer to do.

have the tests every worked correctly? There are things I needed to change in the tests in order to get them to run that had nothing to do with any of the changes I made to the library.