plot: changed conditions for replotting of same data
A user might want to show some values in the first subplot and all the rest in a second subplot:
$ cat log | cantools plot my.dbc '*device1.current' '*device2.current' - 'current'
Here the data should not be replotted. (But has been before this commit.)

On the other hand a user might want to emphasize the data points in a different color
to see how much time is between two messages:
$ cat log | cantools plot my.dbc '*device1.current:b-' '*device1.current:ro'
In this case the data should be replotted. (But was not before this commit.)