plot test python3.6 compatibility
datetime.fromisoformat was defined in Python 3.7.

In the last commit I have replaced datetime.fromisoformat by datetime.strptime in the system under test.
After that I have checked that the tests still pass before copying the strptime format
from the system under test to the test itself in this commit to ensure that the format is correct.