subplot args implemented --title and --ylabel
this is a continuation of commit 8970d75758e0a5663fbab0234076bf4d2a992126
--xlabel, --ymin and --ymax are still not implemented

I have switched from parse_known_args to parse_args
so that misspelled subplot arguments are not misinterpreted as signal names.
Therefore I had to add a positional argument "signals"
which I need to strip from the help in line 557.