bugfix plot --title
bugfix for commit a3c36bd836017888025523cdc33e342154c87a69
plt.title works differently than expected.
Instead of setting a title for the entire figure it sets a title for the last subplot.
I have failed to find an option to set a title for the entire figure.
Instead it sets the title of the first (top most) subplot.