plot: automatic ylabels
This is important since commit ca60f98fb033d3ddaaf7135e890c24d0d28cdab9
to see which which signals belong to which axis.

This broke most tests because now there is an unexpected call to subplot.set.
I have solved that by adding an argument to the constructor of SubplotMock
so that calls to subplot.set are usually ignored
but can be checked in some tests where this is relevant.

I am overriding _get_child_mock because otherwise child mocks would be of the same type like the parent
and since I am overriding the constructor it takes different arguments.
The doc string of that method says explicitly "Subclasses of Mock may want to override this".