README.rst: call the 'cantools' and 'pip' commands prefixed by python3
This ought to be the same anyway, but at least on my machine, running
`setup.py install` does not install an executable called
`cantools`. The same might be true for `pip`. Note that this only
fixes the instructions for people who use a UNIX-like operating
system, on Windows the name of the python interpreter might be
different. Thanks to [at]eerimoq for the review comments.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Oliver Kopp <oliver.kopp@mbition.io>