python 3.6 compatability
re.escape unnecessarily inserts a backslash before a percent sign in python 3.6.12 but not anymore in python 3.7.3.
Therefore the code worked with python 3.7 but not with python 3.6.
I have solved the issue by first replacing the wildcards with other wildcards which are treated equally in both versions.
Curly braces have a special meaning in regex therefore they should always be escaped.