ARXML: Introduce a ignore list for units
If one of these "units" is encountered, it shall be treated as if
there was no unit present. This is a hack around specifying "pseudo
units" which are required by some tools. I'm aware that this currently
only works if the display name of the pseudo unit is "NoUnit", but all
of the files which I have to use on a daily basis use this. Also,
patches for adding other conventions should be rather trivial...

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Oliver Kopp <oliver.kopp@mbition.io>