CLI: make loading of subparses more bullet-proof
After this, if a given subparser is not available for whatever reason,
the CLI still works with the remaining subparsers and an error message
is printed if a non-working subparser is attempted to be invoked.

Note that this patch might slightly complicate debugging subparsers
because it is slightly harder to get hold of the traceback of the
import error. That said, the traceback is currently not printed
out-of-the-box anyway, so there will not be much difference from a
developer perspective...

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Oliver Kopp <oliver.kopp@mbition.io>