rerun LaTeX
not sure how often is needed for ltablex.

First the documentation says "the document should be run through latex twice."
but more prominently it states: "The document needs to be latexed a couple times in general."
which sounds like twice might not be enough. I think with three times I should be safe.