Update control flow in _decode_field
- before the update choices was always applied to the value, if it
succeeded, it would return, but if it failed scaling would be applied. I
believe this behavior to be inconsistent.
- This code passes all, but one test (cantools/tests/test_database.py", line 1266, in test_the_homer_encode_decode_choice_scaling self.assertEqual(decoded, decoded_message)).
- The code has different behavior from the previous code, when a signal
posesses choices, but also scaling. I dont know about the official CAN
spec, but this kind of message should not exist. Either choices or
scaling is used, but not both. To prevent this behavior one could add a
check to the database, which can be run mandatorly or optionally.
- I also reformatted the code with black as I'm not a fan of the \ syntax