separator for multiplexed and multiplexer
if a signal is both multiplexed and a multiplexer
separate the m0 and M by a comma.
I have moved the function call of get_mux
out of the dictionary to avoid it's return value
being processed by texify.
With the current separator of comma and space
that does not make a difference but my first idea
was to use a \, and that should not be processed by texify.