fix tests for infile and outfile on windows
bugfix for commit f3246e4de0e9c92149e01fc7c5c93f3755cd3421

The output of the tests is written to a tmp directory
therefore the definition of \outfile changes with every test
and does not match the golden master.
Therefore I am changing the definition of \outfile
before comparing it to the golden master.
Before this commit I accomplished that with os.path.split.
That works on Linux but not on Windows.

On windows path.sep is "\", a character which is inserted by texify, e.g. "_" -> "\_".
Therefore os.path.split strips too much on Windows.