DBC: include the cycle time of messages when writing files
the cycle time is already parsed, it was just not written: The idea of
this patch is to write attributes which are directly represented by
Signal() or Message() objects without relying on the
`{signal,message}.dbc` sub-object. The rationale for this is that the
user might have loaded a database from a file, then manipulated a few
attributes and then writes back a DBC file. (thus,
`{signal,message}.dbc` may be present or not, and the respective
attribute might be defined and consistent with the rest of the object
or not.)

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Oliver Kopp <oliver.kopp@mbition.io>