ARXML: change the approach to references
Contrary to what I believed in the past, not only packages can contain
sub-elements that can be referenced. We thus change the approach to
ARXML references to precompute them by inspecting the whole file and
simply look them up in the dictionary directly. The approach has the
disadvantage that it requires more memory, but it has the advantages
that references now work for arbitrary nestings of XML elements and
that it is more performant.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Oliver Kopp <oliver.kopp@mbition.io>