ARXML: always pretend to be version 4.0 for AR4
For some reason, all AR 4 revisions always use
"http://autosar.org/schema/r4.0" as their XML namespace. To find out
the exact AUTOSAR revision (i.e., 4.0, 4.1, 4.2, ...), the
"xsi:schemaLocation" attribute of the root tag needs to be
examined. Since this is pretty fragile (the used naming scheme has
changed during the AR4 journey and with the latest naming scheme there
seems to be no programmatic way to associate the value of
schemaLocation with the AR revision), we pretend to always use AR
4.0. If incompatible changes between AR4 revisions are encountered in
the future, it can be supported by implementing all posibilities and
backtracking if necessary.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Oliver Kopp <oliver.kopp@mbition.io>