shorten the named values of signals
This gets rid if redundant prefixes of the named values of signals
thus reducing visual noise when decoding such signals considerably
while still guaranteeing the uniqueness and validity of the new names.

On one hand, this requires minor modifications for some of the user
scripts (sorry in advance to all affected users), but on the other
hand, I think that it improves usability considerably enough to enable
it by default.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Oliver Kopp <oliver.kopp@mbition.io>