ARXML: be more lenient about accepting floating point values
Besides accepting fractional values for floating point signals, we now
also accept floating point values whenever we expect integers provided
that the specified value exhibits a decimal value of `.0`. This is a
superset of the AUTOSAR mandated behavior, i.e. all valid ARXML files
can be loaded correctly but not all files that can be loaded are
valid. (We already do this in some places so we do not really get less
standard-adherent...)

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Oliver Kopp <oliver.kopp@mbition.io>