Allow to specify the pattern for unused bits
The purpose of this is to avoid undefined behavior when encoding
frames that contain bits which are not used by any signal. Currently,
this pattern implicitly is 0xff if encoding with the `padding`
argument set to `True`, and 0x00 otherwise. ARXML allows to specify
something arbitrary, so let's also support this in cantools. IMO, this
is quite a fringe feature and I'm not even completely sure if it is
implemented fully correctly here: I assumed that a single byte gets
specified as an integer (thus, endianness does not matter) and that the
bit pattern is not aligned with the beginning of unspecified sections,
i.e., the encoded message starts out as a sequence of the unspecified
bit pattern bytes and then the parts which define signals get
overwritten by actual data.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Oliver Kopp <oliver.kopp@mbition.io>