bugfix prune_signal_choices: don't shorten signal names which don't include an underscore
The algorithm assumes the names would include an underscore and if they don't it stripped too much.
I have solved that by not stripping names which include no underscore.
Alternatively the problem could have been solved by changing the algorithm
to be able to cope with names without underscores. But the doc string explicitly
talks about underscores so I am assuming this was the intended behavior.