Container Messages: Implement de- and encoding
Since decoding container messages results in a list of
(Message, SignalDict) tuples instead of a single SignalDict,
`decode()` must be explicitly instructed to handle them. The reason
for this is that code which works before this commit will do so
afterwards (if it handles `DecodeError` for `decode()`, which is
basically mandatory anyway), but if container messages were decoded by
default, existing code would almost certainly raise unexpected
exceptions if any container message were encountered.

Besides implementing the actual de- and encoding functionality, this
commit adapts the `decode` and `monitor` subparsers to handle
containers.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Katrin Bauer <katrin.bauer@mbition.io>