fix the tester test for newer python can versions
for python-can 4.0, the objects passed to can.Notifier need to be
derived from can.BusABC where it seems to have been optional
before. Thanks to [at]zariiii9003 for pointing this out!

also, we now specify the 4.0 minimum required python-can version.

this commit was motivated by
https://github.com/cantools/cantools/pull/399#issuecomment-1046026905
but is unrelated to #399 and only coincides with the timing of
python-can 4.0.0.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>