monitor: Improve handling of container messages
Instead of displaying the contents of a container message directly, we
now only give the "table of contents" of the container in a single
line and display the contained messages separately using the naming
scheme "{container_message_name} :: {contained_message_name}". This
approach has the huge advantage of significantly reducing the
likelyhood of jittering the screen content to the point of uselessness
every time a container message is received. (IOW, before this commit,
the `monitor` subparser IMO only exhibits a very limited amount of
usefulness as soon as containers are involved.)

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Katrin Bauer <katrin.bauer@mbition.io>