monitor: Allow the payload of frames to be longer than specified by the respective message
In the case that a CAN frame exhibits extra bytes, it can be correctly
decoded. Since some ECUs tend to do this, I am in favour of dealing
down the strictness a notch in this case.

A possible improvement would be to also display the trailing garbage,
but IMO this is not important...

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Katrin Bauer <katrin.bauer@mbition.io>