add E2E unit tests
Besides adapting the data files, this commit creates a file for
AUTOSAR specific unit tests. So far, it only contains E2E tests, but
e.g., the container stuff could also be moved there in the
future. (The issue with containers is that conceptually they are not a
purely AUTOSAR specific feature...)

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Oliver Kopp <oliver.kopp@mbition.io>