loader: catch any exception from the format loaders
before, only a selected few were caught. This caused quite a bit of
confusion for me, because for these which were caught, the traceback
was mangled, and for those which were not, the traceback was
okay. With this, it should now always work and it can be argued that
it is even more correct because a file of a given format is
unsupported regardless of what exception its format loader raises
during a loading attempt...

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Gerrit Ecke <gerrit.ecke@mbition.io>