Apply scaling to initial values when scaling is requested
Scaling was always applied to the initial value even if scaling
was passed as False to the Tester. The values are always of the
type decimal.Decimal. This meant that all initial values were
handeled as string in the _encode_fields function since the check
only looked for floats and ints.

Please note that this commit implies that the type decimal.Decimal
will be used for all signals with an initial value when scaling is
True.