list: print units for minimum, maximum, initial and invalid values
If applicable, we now print the unit for the pre-defined signal values
(because these are physical values) and we also alow to specify the
format specifier for physical signal values in the list parser. The
latter change allows to get around rounding errors when comparing
databases stemming from conversion of one file format to another.

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Gerrit Ecke <gerrit.ecke@mbition.io>