Remove usage of Decimal objects
The issue with the use of Decimal objects to avoid rounding errors is
that most people (me included) were not very confident in its usage
which made cantools (IMO) disproportionately harder to modify and it
lead to lurking bugs as well as maintenance and performance issues if
the decimal number objects were used for computations instead of the
"normal" ones.

The most difficult drawback of this patch is that DBC files cannot be
compared byte-for-byte anymore because rounding errors can be
introduced during a load-store round-trip. What we do instead is
ensure that "round tripped" file is "close enough" to the original
one. Besides getting around the issue that the two database files
might not be 100% identical due to rounding errors, this approach also
allows to compare databases stemming from different formats (with
minor modifications to avoid comparing the file format specific
information).

Unfortunately, with this patch the output of the C code generator
changes slightly, causing some code churn...

Signed-off-by: Andreas Lauser <andreas.lauser@mbition.io>
Signed-off-by: Gerrit Ecke <gerrit.ecke@mbition.io>