reject DBCs with unused VAL_ choices
I came across a DBC in the wild which contained a VAL_ line that specified
the wrong frame ID for its signal.  cantools accepted the DBC file,
but failed to interpret the signal correctly.

This commit tries to detect this DBC error by removing VAL_ choices
as they are used by Messages, and raising an exception if any choices
remain after parsing the entire DBC.