reject DBCs with unused SIG_VALTYPE_ entries
I came across a DBC in the wild which contained a SIG_VALTYPE_ line that
specified the wrong frame ID for its signal.  cantools accepted the DBC
file, but failed to interpret the signal correctly.

This commit tries to detect this DBC error by removing SIG_VALTYPE_
choices from the signal_types dict as they are used by Messages, and
raising an exception if any entries remain after parsing the entire DBC.